Update in flutter, remove -  as int of the codes

+++ pratic/dart/exs/ex011/main.dart
@@ -9,7 +9,7 @@ void main(List<String> args) {
 calcDps(List<int> dps) {
   int total = 0;
   for (int i = 0; i < dps.length; i++) {
-    total += dps[i] as int;
+    total += dps[i];
   }
 
   print("Minha despesa e: $total");
